requirements.txt
requests ~= 2.0
Django ~= 4.0
Markdown ~= 3.4
django-admin-sortable2 ~= 2.1
django-cleanup ~= 7.0
django-allauth ~= 0.54
easy-thumbnails ~= 2.8
Pillow ~= 9.0
pyphen ~= 0.14
